# rest-server/src/IO.Swagger/wwwroot/swagger-original.json
{
  "openapi" : "3.0.1",
  "info" : {
    "title" : "Paperless Rest Server",
    "version" : "1.0"
  },
  "servers" : [ {
    "url" : "/"
  } ],
  "paths" : {
    "/api/correspondents" : {
      "get" : {
        "tags" : [ "Correspondents" ],
        "operationId" : "GetCorrespondents",
        "parameters" : [ {
          "name" : "page",
          "in" : "query",
          "required" : false,
          "style" : "form",
          "explode" : true,
          "schema" : {
            "type" : "integer",
            "format" : "int32"
          }
        }, {
          "name" : "full_perms",
          "in" : "query",
          "required" : false,
          "style" : "form",
          "explode" : true,
          "schema" : {
            "type" : "boolean"
          }
        } ],
        "responses" : {
          "200" : {
            "description" : "Success",
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/inline_response_200"
                },
                "examples" : {
                  "default" : {
                    "value" : "{\"count\":2,\"next\":null,\"previous\":null,\"all\":[2,1],\"results\":[{\"id\":2,\"slug\":\"edcorr2\",\"name\":\"edcorr2\",\"match\":\"\",\"matching_algorithm\":0,\"is_insensitive\":true,\"document_count\":0,\"last_correspondence\":null,\"owner\":3,\"permissions\":{\"view\":{\"users\":[],\"groups\":[]},\"change\":{\"users\":[],\"groups\":[]}}},{\"id\":1,\"slug\":\"edtest\",\"name\":\"edtest\",\"match\":\"\",\"matching_algorithm\":6,\"is_insensitive\":true,\"document_count\":1,\"last_correspondence\":\"2023-09-08T00:00:00Z\",\"owner\":3,\"permissions\":{\"view\":{\"users\":[],\"groups\":[]},\"change\":{\"users\":[],\"groups\":[]}}}]}"
                  }
                }
              }
            }
          }
        }
      },
      "post" : {
        "tags" : [ "Correspondents" ],
        "operationId" : "CreateCorrespondent",
        "requestBody" : {
          "content" : {
            "application/json" : {
              "schema" : {
                "$ref" : "#/components/schemas/api_correspondents_body"
              },
              "examples" : {
                "default" : {
                  "value" : "{\"name\":\"edcorr2\",\"matching_algorithm\":0,\"match\":\"\",\"is_insensitive\":true,\"owner\":3}"
                }
              }
            }
          }
        },
        "responses" : {
          "200" : {
            "description" : "Success",
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/api_correspondents_body"
                },
                "examples" : {
                  "default" : {
                    "value" : "{\"id\":2,\"slug\":\"edcorr2\",\"name\":\"edcorr2\",\"match\":\"\",\"matching_algorithm\":0,\"is_insensitive\":true,\"owner\":3,\"user_can_change\":true}"
                  }
                }
              }
            }
          }
        }
      }
    },
    "/api/correspondents/{id}" : {
      "put" : {
        "tags" : [ "Correspondents" ],
        "operationId" : "UpdateCorrespondent",
        "parameters" : [ {
          "name" : "id",
          "in" : "path",
          "required" : true,
          "style" : "simple",
          "explode" : false,
          "schema" : {
            "type" : "integer",
            "format" : "int32"
          }
        } ],
        "requestBody" : {
          "content" : {
            "application/json" : {
              "schema" : {
                "$ref" : "#/components/schemas/correspondents_id_body"
              },
              "examples" : {
                "default" : {
                  "value" : "{\"id\":2,\"slug\":\"edcorr2\",\"name\":\"edcorr2\",\"match\":\"corr\",\"matching_algorithm\":3,\"is_insensitive\":true,\"document_count\":0,\"last_correspondence\":null,\"owner\":3,\"permissions\":{\"view\":{\"users\":[],\"groups\":[]},\"change\":{\"users\":[],\"groups\":[]}},\"set_permissions\":{\"view\":{\"users\":[],\"groups\":[]},\"change\":{\"users\":[],\"groups\":[]}},\"permissions_form\":{\"owner\":3,\"set_permissions\":{\"view\":{\"users\":[],\"groups\":[]},\"change\":{\"users\":[],\"groups\":[]}}}}"
                }
              }
            }
          }
        },
        "responses" : {
          "200" : {
            "description" : "Success",
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/inline_response_200_1"
                },
                "examples" : {
                  "default" : {
                    "value" : "{\"id\":2,\"slug\":\"edcorr2\",\"name\":\"edcorr2\",\"match\":\"corr\",\"matching_algorithm\":3,\"is_insensitive\":true,\"document_count\":0,\"last_correspondence\":null,\"owner\":3,\"user_can_change\":true}"
                  }
                }
              }
            }
          }
        }
      },
      "delete" : {
        "tags" : [ "Correspondents" ],
        "operationId" : "DeleteCorrespondent",
        "parameters" : [ {
          "name" : "id",
          "in" : "path",
          "required" : true,
          "style" : "simple",
          "explode" : false,
          "schema" : {
            "type" : "integer",
            "format" : "int32"
          }
        } ],
        "responses" : {
          "204" : {
            "description" : "Success"
          }
        }
      }
    },
...
  },
  "components" : {
    "schemas" : {
...
      "inline_response_200" : {
        "required" : [ "all", "count", "next", "previous", "results" ],
        "type" : "object",
        "properties" : {
          "count" : {
            "type" : "integer"
          },
          "next" : {
            "type" : "integer"
          },
          "previous" : {
            "type" : "integer"
          },
          "all" : {
            "type" : "array",
            "items" : {
              "type" : "integer"
            }
          },
          "results" : {
            "type" : "array",
            "items" : {
              "$ref" : "#/components/schemas/inline_response_200_results"
            }
          }
        },
        "example" : {
          "next" : 6,
          "all" : [ 5, 5 ],
          "previous" : 1,
          "count" : 0,
          "results" : [ {
            "owner" : 9,
            "matching_algorithm" : 2,
            "document_count" : 7,
            "is_insensitive" : true,
            "permissions" : {
              "view" : {
                "groups" : [ "", "" ],
                "users" : [ "", "" ]
              },
              "change" : null
            },
            "name" : "name",
            "match" : "match",
            "id" : 5,
            "last_correspondence" : "last_correspondence",
            "slug" : "slug"
          }, {
            "owner" : 9,
            "matching_algorithm" : 2,
            "document_count" : 7,
            "is_insensitive" : true,
            "permissions" : {
              "view" : {
                "groups" : [ "", "" ],
                "users" : [ "", "" ]
              },
              "change" : null
            },
            "name" : "name",
            "match" : "match",
            "id" : 5,
            "last_correspondence" : "last_correspondence",
            "slug" : "slug"
          } ]
        }
      },
      "api_correspondents_body" : {
        "required" : [ "is_insensitive", "match", "matching_algorithm", "name", "owner" ],
        "type" : "object",
        "properties" : {
          "name" : {
            "type" : "string"
          },
          "matching_algorithm" : {
            "type" : "integer"
          },
          "match" : {
            "type" : "string"
          },
          "is_insensitive" : {
            "type" : "boolean"
          },
          "owner" : {
            "type" : "integer"
          }
        },
        "example" : {
          "owner" : 6,
          "matching_algorithm" : 0,
          "is_insensitive" : true,
          "name" : "name",
          "match" : "match"
        }
      },
      "apicorrespondentsid_permissions_view" : {
        "required" : [ "groups", "users" ],
        "type" : "object",
        "properties" : {
          "users" : {
            "type" : "array",
            "items" : { }
          },
          "groups" : {
            "type" : "array",
            "items" : { }
          }
        },
        "example" : {
          "groups" : [ "", "" ],
          "users" : [ "", "" ]
        }
      },
      "apicorrespondentsid_permissions" : {
        "required" : [ "change", "view" ],
        "type" : "object",
        "properties" : {
          "view" : {
            "$ref" : "#/components/schemas/apicorrespondentsid_permissions_view"
          },
          "change" : {
            "$ref" : "#/components/schemas/apicorrespondentsid_permissions_view"
          }
        },
        "example" : {
          "view" : {
            "groups" : [ "", "" ],
            "users" : [ "", "" ]
          },
          "change" : null
        }
      },
      "apicorrespondentsid_permissions_form" : {
        "required" : [ "owner", "set_permissions" ],
        "type" : "object",
        "properties" : {
          "owner" : {
            "type" : "integer"
          },
          "set_permissions" : {
            "$ref" : "#/components/schemas/apicorrespondentsid_permissions"
          }
        }
      },
      "correspondents_id_body" : {
        "required" : [ "document_count", "id", "is_insensitive", "last_correspondence", "match", "matching_algorithm", "name", "owner", "permissions", "permissions_form", "set_permissions", "slug" ],
        "type" : "object",
        "properties" : {
          "id" : {
            "type" : "integer"
          },
          "slug" : {
            "type" : "string"
          },
          "name" : {
            "type" : "string"
          },
          "match" : {
            "type" : "string"
          },
          "matching_algorithm" : {
            "type" : "integer"
          },
          "is_insensitive" : {
            "type" : "boolean"
          },
          "document_count" : {
            "type" : "integer"
          },
          "last_correspondence" : {
            "type" : "integer"
          },
          "owner" : {
            "type" : "integer"
          },
          "permissions" : {
            "$ref" : "#/components/schemas/apicorrespondentsid_permissions"
          },
          "set_permissions" : {
            "$ref" : "#/components/schemas/apicorrespondentsid_permissions"
          },
          "permissions_form" : {
            "$ref" : "#/components/schemas/apicorrespondentsid_permissions_form"
          }
        }
      },
      "inline_response_200_1" : {
        "required" : [ "document_count", "id", "is_insensitive", "last_correspondence", "match", "matching_algorithm", "name", "owner", "slug", "user_can_change" ],
        "type" : "object",
        "properties" : {
          "id" : {
            "type" : "integer"
          },
          "slug" : {
            "type" : "string"
          },
          "name" : {
            "type" : "string"
          },
          "match" : {
            "type" : "string"
          },
          "matching_algorithm" : {
            "type" : "integer"
          },
          "is_insensitive" : {
            "type" : "boolean"
          },
          "document_count" : {
            "type" : "integer"
          },
          "last_correspondence" : {
            "type" : "integer"
          },
          "owner" : {
            "type" : "integer"
          },
          "user_can_change" : {
            "type" : "boolean"
          }
        },
        "example" : {
          "owner" : 5,
          "matching_algorithm" : 6,
          "user_can_change" : true,
          "document_count" : 1,
          "is_insensitive" : true,
          "name" : "name",
          "match" : "match",
          "id" : 0,
          "last_correspondence" : 5,
          "slug" : "slug"
        }
      },
...
      "inline_response_200_results" : {
        "required" : [ "document_count", "id", "is_insensitive", "last_correspondence", "match", "matching_algorithm", "name", "owner", "permissions", "slug" ],
        "type" : "object",
        "properties" : {
          "id" : {
            "type" : "integer"
          },
          "slug" : {
            "type" : "string"
          },
          "name" : {
            "type" : "string"
          },
          "match" : {
            "type" : "string"
          },
          "matching_algorithm" : {
            "type" : "integer"
          },
          "is_insensitive" : {
            "type" : "boolean"
          },
          "document_count" : {
            "type" : "integer"
          },
          "last_correspondence" : {
            "type" : "string"
          },
          "owner" : {
            "type" : "integer"
          },
          "permissions" : {
            "$ref" : "#/components/schemas/apicorrespondentsid_permissions"
          }
        },
        "example" : {
          "owner" : 9,
          "matching_algorithm" : 2,
          "document_count" : 7,
          "is_insensitive" : true,
          "permissions" : {
            "view" : {
              "groups" : [ "", "" ],
              "users" : [ "", "" ]
            },
            "change" : null
          },
          "name" : "name",
          "match" : "match",
          "id" : 5,
          "last_correspondence" : "last_correspondence",
          "slug" : "slug"
        }
      },
...
    }
  }
}